feat(example): add httproute, RBAC and Gateway API resources for Kong
This commit introduces necessary RBAC permissions for the Kong service account to manage Gateway and HTTPRoute resources within the 'prj-example-1' namespace. It includes the creation of Roles and RoleBindings to ensure the Kong Ingress Controller can update the status of HTTPRoute resources and manage Gateway configurations effectively. Additionally, this commit adds the Gateway and HTTPRoute resources to enable routing configurations through Kong for the 'prj-example-1' frontend service.

Signed-off-by: Ales Verbic <verbotenj@blinklabs.io>