refactor(kong): remove 'depends-on helm' as it is non-functional (#157)
Currently, during terraform plan, the CRDs are not deployed yet, leading to a failure. As a workaround, we target the helm_release first by running terraform apply -target=helm_release.kong. This ensures the Helm release is applied before the CRDs, preventing the deployment failure. This is a temporary solution until a more permanent fix is implemented.

Signed-off-by: Ales Verbic <verbotenj@blinklabs.io>